Be/dev

+++ BE/src/main/java/shop/petmily/domain/pet/service/PetService.java
@@ -1,10 +1,13 @@
 package shop.petmily.domain.pet.service;
 
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import shop.petmily.domain.pet.entity.Pet;
 import shop.petmily.domain.pet.repository.PetRepository;
 import shop.petmily.global.AWS.service.S3UploadService;
+import shop.petmily.global.exception.BusinessLogicException;
+import shop.petmily.global.exception.ExceptionCode;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
@@ -37,13 +40,13 @@ public class PetService {
     public Pet updatePet(Pet pet, MultipartFile file) throws IOException {
         Pet verifiedPet = verifiedPet(pet.getPetId());
 
-        Optional.ofNullable(pet.getAge())
-                .ifPresent(age -> verifiedPet.setAge(age));
-        Optional.ofNullable(pet.getWeight())
-                .ifPresent(weight -> verifiedPet.setWeight(weight));
+        verifiedPetOwner(verifiedPet.getMember().getMemberId(), pet.getMember().getMemberId());
+
+        if(pet.getAge() != 0) verifiedPet.setAge(pet.getAge());
+        if(pet.getWeight() != 0) verifiedPet.setWeight(pet.getWeight());
 
         if(file != null) {
-            uploadService.deleteFile(verifiedPet.getPhoto());
+            if(verifiedPet.getPhoto() != null) uploadService.deleteFile(verifiedPet.getPhoto());
             verifiedPet.setPhoto(uploadService.saveFile(file));
         }
 
@@ -52,13 +55,35 @@ public class PetService {
         return repository.save(verifiedPet);
     }
 
-    public void deletePet(long petId){
-        repository.delete(verifiedPet(petId));
+    public Pet photoDelete(long petId, Long requestMemberId){
+        Pet verifiedPet = verifiedPet(petId);
+        verifiedPetOwner(verifiedPet.getMember().getMemberId(), requestMemberId);
+
+        uploadService.deleteFile(verifiedPet.getPhoto());
+        verifiedPet.setPhoto(null);
+        verifiedPet.setLastModifiedAt(LocalDateTime.now());
+
+        return repository.save(verifiedPet);
+    }
+
+    public Pet findPet(long petId){
+        Pet verifiedPet = verifiedPet(petId);
+        return  verifiedPet;
+    }
+
+    public void deletePet(long petId, Long requestMemberId){
+        Pet verifiedPet = verifiedPet(petId);
+        verifiedPetOwner(verifiedPet.getMember().getMemberId(), requestMemberId);
+        repository.delete(verifiedPet);
     }
 
     private Pet verifiedPet(long petId) {
         Optional<Pet> optionalPet = repository.findById(petId);
-        Pet pet = optionalPet.orElseThrow(() -> new RuntimeException());
+        Pet pet = optionalPet.orElseThrow(() -> new BusinessLogicException(ExceptionCode.PET_NOT_EXIST));
         return pet;
+    }
+
+    private void verifiedPetOwner(long originMemberId, Long requestMemberId) {
+        if (originMemberId != requestMemberId) throw new BusinessLogicException(ExceptionCode.NOT_ALLOW_MEMBER);
     }
 }
